auth_with_self expects modinfo to be sent, fixed bug in GetMemberSize
